Updated routing to use query params
