+++ sites/partners/netlify.toml
@@ -13,8 +13,3 @@ GTM_KEY = "GTM-TBD"
 
 [context.branch-deploy.environment]
 GTM_KEY = "GTM-TBD"
-
-[[redirects]]
-  from = "/*"
-  to = "/index.html"
-  status = 200
