+++ sites/partners/next.config.js
@@ -64,14 +64,6 @@ module.exports = withCSS(
 
           return languagePaths
         },
-        rewrites() {
-          return [
-            {
-              source: "/:path*",
-              destination: "/",
-            },
-          ]
-        },
       })
     )
   )
