+++ sites/partners/pages/application/index.tsx
@@ -10,24 +10,24 @@ import {
   ApiClientContext,
   AlertBox,
 } from "@bloom-housing/ui-components"
-import { useSingleApplicationData } from "../../../lib/hooks"
-import Layout from "../../../layouts/application"
+import { useSingleApplicationData } from "../../lib/hooks"
+import Layout from "../../layouts/application"
 import { ApplicationStatus } from "@bloom-housing/backend-core/types"
 import {
   DetailsMemberDrawer,
   MembersDrawer,
-} from "../../../src/applications/PaperApplicationDetails/DetailsMemberDrawer"
+} from "../../src/applications/PaperApplicationDetails/DetailsMemberDrawer"
 
-import { DetailsApplicationContext } from "../../../src/applications/PaperApplicationDetails/DetailsApplicationContext"
-import { DetailsApplicationData } from "../../../src/applications/PaperApplicationDetails/sections/DetailsApplicationData"
-import { DetailsPrimaryApplicant } from "../../../src/applications/PaperApplicationDetails/sections/DetailsPrimaryApplicant"
-import { DetailsAlternateContact } from "../../../src/applications/PaperApplicationDetails/sections/DetailsAlternateContact"
-import { DetailsHouseholdMembers } from "../../../src/applications/PaperApplicationDetails/sections/DetailsHouseholdMembers"
-import { DetailsHouseholdDetails } from "../../../src/applications/PaperApplicationDetails/sections/DetailsHouseholdDetails"
-import { DetailsPreferences } from "../../../src/applications/PaperApplicationDetails/sections/DetailsPreferences"
-import { DetailsHouseholdIncome } from "../../../src/applications/PaperApplicationDetails/sections/DetailsHouseholdIncome"
-import { DetailsTerms } from "../../../src/applications/PaperApplicationDetails/sections/DetailsTerms"
-import { DetailsAside } from "../../../src/applications/PaperApplicationDetails/DetailsAside"
+import { DetailsApplicationContext } from "../../src/applications/PaperApplicationDetails/DetailsApplicationContext"
+import { DetailsApplicationData } from "../../src/applications/PaperApplicationDetails/sections/DetailsApplicationData"
+import { DetailsPrimaryApplicant } from "../../src/applications/PaperApplicationDetails/sections/DetailsPrimaryApplicant"
+import { DetailsAlternateContact } from "../../src/applications/PaperApplicationDetails/sections/DetailsAlternateContact"
+import { DetailsHouseholdMembers } from "../../src/applications/PaperApplicationDetails/sections/DetailsHouseholdMembers"
+import { DetailsHouseholdDetails } from "../../src/applications/PaperApplicationDetails/sections/DetailsHouseholdDetails"
+import { DetailsPreferences } from "../../src/applications/PaperApplicationDetails/sections/DetailsPreferences"
+import { DetailsHouseholdIncome } from "../../src/applications/PaperApplicationDetails/sections/DetailsHouseholdIncome"
+import { DetailsTerms } from "../../src/applications/PaperApplicationDetails/sections/DetailsTerms"
+import { DetailsAside } from "../../src/applications/PaperApplicationDetails/DetailsAside"
 
 export default function ApplicationsList() {
   const router = useRouter()
@@ -42,7 +42,7 @@ export default function ApplicationsList() {
   async function deleteApplication() {
     try {
       await applicationsService.delete({ applicationId })
-      void router.push(`/listings/${application?.listing?.id}/applications`)
+      void router.push(`/listings/applications?listing=${application?.listing?.id}`)
     } catch (err) {
       setErrorAlert(true)
     }
