+++ sites/partners/pages/listings/applications/add.tsx
@@ -1,6 +1,5 @@
 import React from "react"
 import Head from "next/head"
-import { useRouter } from "next/router"
 import {
   AppearanceStyleType,
   Button,
@@ -9,21 +8,16 @@ import {
   StatusBar,
   t,
 } from "@bloom-housing/ui-components"
-import { ApplicationStatus } from "@bloom-housing/backend-core/types"
 import Layout from "../../../layouts/application"
 import PaperApplicationForm from "../../../src/applications/PaperApplicationForm/PaperApplicationForm"
-import { useSingleApplicationData } from "../../../lib/hooks"
+import { useRouter } from "next/router"
 
 const NewApplication = () => {
   const metaDescription = ""
   const metaImage = "" // TODO: replace with hero image
 
   const router = useRouter()
-  const applicationId = router.query.id as string
-
-  const { application } = useSingleApplicationData(applicationId)
-
-  if (!application) return false
+  const listingId = router.query.listing as string
 
   return (
     <Layout>
@@ -39,18 +33,11 @@ const NewApplication = () => {
             {t("t.back")}
           </Button>
         }
-        tagStyle={
-          application?.status == ApplicationStatus.submitted
-            ? AppearanceStyleType.success
-            : AppearanceStyleType.primary
-        }
-        tagLabel={
-          application?.status
-            ? t(`application.details.applicationStatus.${application.status}`)
-            : t(`application.details.applicationStatus.submitted`)
-        }
+        tagStyle={AppearanceStyleType.primary}
+        tagLabel={t(`application.details.applicationStatus.draft`)}
       />
-      <PaperApplicationForm listingId={application.listing.id} />
+
+      <PaperApplicationForm listingId={listingId} />
     </Layout>
   )
 }
